consolidate directus data fetching
